tab2 is almost fully functional, except the search bar.
created 2 few tables for in-forced policies and reports.

In-forced policies tab(tab3) does work as it is intended, but it needs logic improvements. I think policies should be tied somehow to recipients.

Tab4 reports is somewhat functional. Reports are stored in a table in the database and then pulled from there to the tab4 tableView. However, there is no way to add to the database through the program.